requirements.txt
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
distlib==0.3.6
filelock==3.12.2
identify==2.5.24
idna==3.4
multimethod==1.9.1
nodeenv==1.8.0
platformdirs==3.5.3
pre-commit==3.3.3
pydantic==1.10.9
python-decouple==3.8
PyYAML==6.0
pytest==7.3.2
requests==2.31.0
six==1.16.0
typing-extensions==4.6.3
urllib3==2.0.3
virtualenv==20.23.0
